Move .encoding_cleanup to Utils
