+++ lib/utils.rb
@@ -4,6 +4,11 @@ require 'logger'
 require_relative '../config/boot'
 
 module Utils
+  def self.encoding_cleanup(row)
+    # cleans up cyrlic encoding i︠a︡ to i͡a
+    row.gsub(/\ufe20(.{1,2})\ufe21/, "\u0361\\1")
+  end
+
   def self.balance_parentheses(string)
     open_deletes = []
     close_deletes = []
